backend/profile-service/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
motor>=3.6.0
pymongo>=4.7.0
redis>=5.0.0
strawberry-graphql[fastapi]>=0.243.0
firebase-admin>=6.5.0
python-multipart>=0.0.9
